adds trader, various improvements

+++ code/z_adventurezones/spacevegas.dm
@@ -14,25 +14,37 @@
 
 /obj/structure/brokentaxipod
 	name = "Broken Taxi Pod"
-	desc = "A small one-person pod that appears to automatically take whoever's in it somewhere. These must've gone out of use years ago. Seems like it could use some new parts."
+	desc = "A small one-person pod that automatically takes whoever's in it somewhere. These must've gone out of use years ago. Seems to be broken and rusty."
 	icon = 'icons/obj/ship.dmi'
 	icon_state = "escape"
 	anchored = 1
 	dir = 4
+	var/progress = 0
 
 	attackby(obj/item/W as obj, mob/living/user as mob)
-		var/obj/item/taxipodengine/C = W
-		if (istype(C))
-			boutput(user, "You replace the pod's rusty engine.")
-			new /obj/machinery/vehicle/escape_pod/taxipod(src.loc)
-			new /obj/item/taxipodengine/rusty(src.loc)
-			user.u_equip(W)
-			qdel(W)
-			qdel(src)
+		switch (progress)
+			if (0)
+				if (isweldingtool(W) && W:try_weld(user,0))
+					boutput(user, "You significantly repair the pod.")
+					progress++
+			if (1)
+				if (ispryingtool(W))
+					boutput(user, "You remove the pod's rusty engine.")
+					playsound(src.loc, "sound/items/Crowbar.ogg", 50, 1)
+					new /obj/item/taxipodengine/rusty(src.loc)
+					progress++
+			if (2)
+				if (istype(W, /obj/item/taxipodengine))
+					boutput(user, "You replace the pod's rusty engine.")
+					playsound(src.loc, "sound/items/Deconstruct.ogg", 50, 0)
+					new /obj/machinery/vehicle/escape_pod/taxipod(src.loc)
+					user.u_equip(W)
+					qdel(W)
+					qdel(src)
 
 /obj/machinery/vehicle/escape_pod/taxipod
 	name = "Taxi Pod V-"
-	desc = "A small one-person pod that appears to automatically take whoever's in it somewhere. These must've gone out of use years ago."
+	desc = "A small one-person pod that automatically takes whoever's in it somewhere. These must've gone out of use years ago."
 	icon = 'icons/obj/ship.dmi'
 	icon_state = "escape"
 	capacity = 1
@@ -40,7 +52,6 @@
 	maxhealth = 100
 	speed = 5
 	anchored = 1
-	dir = 4 // todo: not this. also todo: redo pretty much the entire debris field shuttle, have this thing generate a "wormhole portal" that works like a real portal that doesn't disappear so i don't have to do a shitty taxi system for 9 years
 	var/target = null // goes to the pod entry location (in space vegas)
 
 	New()
@@ -68,23 +79,18 @@
 		return
 
 	escape()
-		if(!launched)
+		if (!launched)
 			launched = 1
 			anchored = 0
-			var/opened_door = 0
-			var/turf_in_front = get_step(src,src.dir)
 
-			for(var/obj/machinery/door/poddoor/D in turf_in_front) // open the door
-				D.open()
-				opened_door = 1
-			if(opened_door) sleep(2 SECONDS) // make sure it's fully open
+			sleep(2 SECONDS)
 
-			var/obj/warp_portal/P = new /obj/warp_portal( src.loc ) // make a portal
+			var/obj/warp_portal/P = new /obj/warp_portal( src.loc ) // make a portal nearby
 			P.transform = matrix(0, MATRIX_SCALE)
 			for(var/i=0, i<2, i++)
 				step(P, src.dir)
 			var/dist = get_dist(src, P)
-			P.pixel_x = -dist*32
+			src.engine.portal_px_offset(P, src.dir, dist)
 			animate(P, transform = matrix(1, MATRIX_SCALE), pixel_x = 0, pixel_y = 0, time = 30, easing = ELASTIC_EASING )
 			sleep(30)
 			P.target = target // make it go to space vegas
@@ -94,12 +100,21 @@
 			sleep(0.5 SECONDS)
 			playsound(src.loc, "sound/effects/flameswoosh.ogg", 100, 0)
 
+			var/hasturned = 0
 			while(!failing)
 				var/loc = src.loc
+				if(hasturned == 0)
+					var/area/spacevegas/area = get_area(src)
+					if (istype(area))
+						hasturned = 1
+						src.dir = 4;
 				step(src,src.dir)
 				if(src.loc == loc) // we hit something
 					var/obj/hit = get_step(src,src.dir)
-					if (istype(hit, /turf/cordon))
+					if (istype(hit, /turf/cordon) || istype(hit, /turf/unsimulated/wall/solidcolor/white))
+						if (src.passengers) // we don't want to be deleting anybody
+							while(src.passengers > 0)
+								src.pilot.set_loc(P.loc) // set_loc calls eject on the pod which decrements passengers
 						qdel(src)
 						break
 					else if (hit != P)
@@ -125,15 +140,17 @@
 		var/obj/machinery/vehicle/P = triggerer
 		if(!istype(P))
 			return
+		if(!P.passengers)
+			return
 		while(P.passengers > 0)
-			P.pilot.set_loc(P.loc) // set_loc calls eject on the pod
+			P.pilot.set_loc(P.loc) // set_loc calls eject on the pod which decrements passengers
 
 
 //////////////////////////////// inside of casino shit
 
 /area/spacevegas
 	name = "Space Casino"
-	icon_state = "green"
+	icon_state = "yellow"
 	force_fullbright = 1
 	sound_environment = 0
 	skip_sims = 0
@@ -145,7 +162,7 @@
 	desc = "A special type of gambling machine that somehow makes items instead of cash."
 	icon = 'icons/misc/casino.dmi'
 	icon_state = "slots-off"
-	//mats = 40
+	mats = 50
 	deconstruct_flags = DECON_NONE
 
 	var/list/junktier = list( // junk tier, 60% chance
@@ -241,19 +258,24 @@
 					var/win_sound = "sound/machines/ping.ogg"
 					var/obj/item/P = null
 
-					if (roll > 1) // self destruction, 2% chance -- intentionally made higher than very rare tier so that this isnt just awesome free items every time all the time
-						playsound(get_turf(src), "sound/misc/airraid_loop_short.ogg", 55, 1) // todo: sprites, particles, effects for this
+					if (roll <= 10) // self destruction, 2% chance -- intentionally made higher than very rare tier so that this isnt just awesome free items every time all the time
+						src.icon_state = "slots-malf"
 						playsound(get_turf(src), "sound/misc/klaxon.ogg", 55, 1)
 						src.visible_message("<span class='subtle'><b>[src]</b> says, 'WINNER! WINNER! JACKPOT! WINNER! JACKPOT! BIG WINNER! BIG WINNER!'</span>")
 						playsound(src.loc, "sound/impact_sounds/Metal_Clang_1.ogg", 60, 1, pitch = 1.2)
 						animate_shake(src,7,5,2)
-						SPAWN_DBG(3.5 SECONDS)
-							src.visible_message("<span class='subtle'><b>[src]</b> says, 'BIG WINNER! BIG WINNER!'</span>")
-							playsound(src.loc, "sound/impact_sounds/Metal_Clang_2.ogg", 60, 1, pitch = 1.5)
-							animate_shake(src,5,7,2)
-							SPAWN_DBG(1.5 SECONDS)
-								qdel(src)
-								return
+						sleep(3.5 SECONDS)
+
+						src.visible_message("<span class='subtle'><b>[src]</b> says, 'BIG WINNER! BIG WINNER!'</span>")
+						playsound(src.loc, "sound/impact_sounds/Metal_Clang_2.ogg", 60, 1, pitch = 0.8)
+						animate_shake(src,5,7,2)
+						sleep(1.5 SECONDS)
+
+						new/obj/decal/implo(src.loc)
+						playsound(src, 'sound/effects/suck.ogg', 60, 1)
+						src.scan.set_loc(src.loc)
+						qdel(src)
+						return
 					else if (roll > 10 && roll <= 15) // very rare tier, 1% chance
 						P = text2path(pick(veryraretier))
 						win_sound = "sound/misc/airraid_loop_short.ogg"
@@ -300,3 +322,110 @@
 /////////// fun exclusive garbage to go with the item slot machine
 
 // todo: sprites for taxipod, taxipod's engine, demagnitized emag; gamblebuddy, barbuddy, making chefbot emaggable, probably some other stuff idk
+
+/////////// HOTDOG GUY TRADER
+
+/datum/commodity/hotdog
+	comname = "Hotdog"
+	comtype = /obj/item/reagent_containers/food/snacks/hotdog
+	desc = "Come get your hotdogs!"
+	price = 20
+	baseprice = 10
+	upperfluc = 15
+	lowerfluc = -10
+
+/datum/commodity/corndog
+	comname = "Corndog"
+	comtype = /obj/item/reagent_containers/food/snacks/corndog
+	desc = "We do corndogs too!"
+	price = 20
+	baseprice = 10
+	upperfluc = 15
+	lowerfluc = -10
+
+/datum/commodity/hauntdog
+	comname = "Hauntdog"
+	comtype = /obj/critter/hauntdog
+	desc = "Not even sure if these are edible but hey, they're cute!"
+	price = 50
+	baseprice = 50
+	upperfluc = 75
+	lowerfluc = -30
+
+/datum/commodity/hotdogocto
+	comname = "Hotdog Octopus"
+	comtype = /obj/item/reagent_containers/food/snacks/hotdog_octo
+	desc = "Look! I can cut them into funny little shapes!"
+	price = 20
+	baseprice = 10
+	upperfluc = 15
+	lowerfluc = -10
+
+/datum/commodity/hotdogbomb
+	comname = "Hotdog Bomb"
+	comtype = /obj/item/gimmickbomb/hotdog
+	desc = "Go on. Make them see what I have to endure."
+	price = 1000
+	baseprice = 1000
+	upperfluc = 1500
+	lowerfluc = -300
+
+/datum/commodity/breadslice
+	comname = "Slice of bread"
+	comtype = /obj/item/reagent_containers/food/snacks/breadslice
+	desc_buy = "I just want bread. Not for hotdogs. I just want to eat bread again. It's been so long."
+	price = 100
+	baseprice = 100
+	upperfluc = 150
+	lowerfluc = -30
+
+/obj/npc/trader/hotdog
+	icon = 'icons/misc/casino.dmi'
+	icon_state = "hotdog"
+	picture = "hotdog.png"
+	name = "Hotdog Guy"
+	trader_area = "/area/spacevegas"
+	angrynope = "HOTDOG GUY DOESN'T WANT TO TALK."
+	whotext = "ITS ME! HOTDOG GUY!"
+
+	New()
+		..()
+		/////////////////////////////////////////////////////////
+		//// sell list //////////////////////////////////////////
+		/////////////////////////////////////////////////////////
+		src.goods_sell += new /datum/commodity/hotdog(src)
+		src.goods_sell += new /datum/commodity/corndog(src)
+		src.goods_sell += new /datum/commodity/hauntdog(src)
+		src.goods_sell += new /datum/commodity/hotdogocto(src)
+		src.goods_sell += new /datum/commodity/costume/hotdog(src)
+		src.goods_sell += new /datum/commodity/hotdogbomb(src)
+		/////////////////////////////////////////////////////////
+		//// buy list ///////////////////////////////////////////
+		/////////////////////////////////////////////////////////
+		src.goods_buy += new /datum/commodity/breadslice(src)
+		/////////////////////////////////////////////////////////
+
+		greeting= {"HEY! COME GET YOUR HOTDOGS! HOTDOGS HERE! PLEASE BUY A HOTDOG! PLEASE!"}
+
+		portrait_setup = "<img src='[resource("images/traders/[src.picture]")]'><HR><B>[src.name]</B><HR>"
+
+		sell_dialogue = "You want to sell something to me...? Well..."
+
+		buy_dialogue = "PLENTY OF HOTDOGS AND HOTDOG ACCESSORIES TO CHOOSE FROM! WE DON'T UH... HAVE ANY BUNS, SO..."
+
+		successful_purchase_dialogue = list("YES! Enjoy, friend!",
+			"THANK YOU!",
+			"THANKS! Come back any time!")
+
+		failed_sale_dialogue = list("I don't want that.",
+			"What use do I have for this...?")
+
+		successful_sale_dialogue = list("WOW! Thanks!",
+			"You actually got it! Thanks!")
+
+		failed_purchase_dialogue = list("You don't have enough...",
+			"I have to make money somehow!")
+
+		pickupdialogue = "HERE YOU GO!"
+
+		pickupdialoguefailure = "You haven't purchased anything, silly!"
